\name{pcalg-internal}
%
%\alias{checkEdges} not exported
%\alias{lm.cov}
%\alias{adj.check}
%\alias{ancTS}
%\alias{caus.inf}
%\alias{has.new.coll}
%\alias{ci.test}
%\alias{dep.triple}
%\alias{faith.check}
%\alias{find.sink}
%\alias{flipEdges}
%\alias{minUncovCircPath}
%\alias{minUncovPdPath}
%\alias{minDiscrPath}
%\alias{my.SpecialDag}
%\alias{reach}

%\alias{updateList}
%--------------------------------------------------------------------------
% FIXME: *do* document [says Martin]  (((or confirm current status))):
\alias{rfci.vStruc}% <- renamed; TODO (MM)
\alias{allDags}%<- mentioned in other help pages !!
\alias{amat2dag}% in ../tests/test_amat2dag.R
\alias{gaussCItest.fast}% will be documented in next CRAN version
%
\title{Internal Pcalg Functions}
\description{
  Internal functions of package \pkg{pcalg}
}
\usage{
amat2dag(amat)
}
\details{
  These are not to be called by the user.
  %% or we have been too lazy to document
}
\keyword{ internal }
